+++ python/tvm/script/parse/doc.py
@@ -1,1309 +1,173 @@
-# pylint: disable=redefined-outer-name,missing-docstring,invalid-name
-# pylint: disable=useless-super-delegation,redefined-builtin
-# pylint: disable=too-few-public-methods,too-many-arguments
-class AST:
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__()
-        self.lineno = lineno
-        self.col_offset = col_offset
-        self.end_lineno = end_lineno
-        self.end_col_offset = end_col_offset
-
-
-class mod(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Module(mod):
-    _FIELDS = ["body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.body = body
-
-
-class Interactive(mod):
-    _FIELDS = ["body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.body = body
-
-
-class Expression(mod):
-    _FIELDS = ["body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.body = body
-
-
-class stmt(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class FunctionDef(stmt):
-    _FIELDS = [
-        "name",
-        "args",
-        "body",
-        "decorator_list",
-        "returns",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        name,
-        args,
-        body,
-        decorator_list,
-        returns,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.name = name
-        self.args = args
-        self.body = body
-        self.decorator_list = decorator_list
-        self.returns = returns
-
-
-class AsyncFunctionDef(stmt):
-    _FIELDS = [
-        "name",
-        "args",
-        "body",
-        "decorator_list",
-        "returns",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        name,
-        args,
-        body,
-        decorator_list,
-        returns,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.name = name
-        self.args = args
-        self.body = body
-        self.decorator_list = decorator_list
-        self.returns = returns
-
-
-class ClassDef(stmt):
-    _FIELDS = [
-        "name",
-        "bases",
-        "keywords",
-        "body",
-        "decorator_list",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        name,
-        bases,
-        keywords,
-        body,
-        decorator_list,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.name = name
-        self.bases = bases
-        self.keywords = keywords
-        self.body = body
-        self.decorator_list = decorator_list
-
-
-class Return(stmt):
-    _FIELDS = ["value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-
-
-class Delete(stmt):
-    _FIELDS = ["targets", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, targets, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.targets = targets
-
-
-class Assign(stmt):
-    _FIELDS = [
-        "targets",
-        "value",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, targets, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.targets = targets
-        self.value = value
-
-
-class AugAssign(stmt):
-    _FIELDS = [
-        "target",
-        "op",
-        "value",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, target, op, value, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.target = target
-        self.op = op
-        self.value = value
-
-
-class AnnAssign(stmt):
-    _FIELDS = [
-        "target",
-        "annotation",
-        "value",
-        "simple",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        target,
-        annotation,
-        value,
-        simple,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.target = target
-        self.annotation = annotation
-        self.value = value
-        self.simple = simple
-
-
-class For(stmt):
-    _FIELDS = [
-        "target",
-        "iter",
-        "body",
-        "orelse",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, target, iter, body, orelse, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.target = target
-        self.iter = iter
-        self.body = body
-        self.orelse = orelse
-
-
-class AsyncFor(stmt):
-    _FIELDS = [
-        "target",
-        "iter",
-        "body",
-        "orelse",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, target, iter, body, orelse, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.target = target
-        self.iter = iter
-        self.body = body
-        self.orelse = orelse
-
-
-class While(stmt):
-    _FIELDS = [
-        "test",
-        "body",
-        "orelse",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, test, body, orelse, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.test = test
-        self.body = body
-        self.orelse = orelse
-
-
-class If(stmt):
-    _FIELDS = [
-        "test",
-        "body",
-        "orelse",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, test, body, orelse, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.test = test
-        self.body = body
-        self.orelse = orelse
-
-
-class With(stmt):
-    _FIELDS = ["items", "body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, items, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.items = items
-        self.body = body
-
-
-class AsyncWith(stmt):
-    _FIELDS = ["items", "body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, items, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.items = items
-        self.body = body
-
-
-class Raise(stmt):
-    _FIELDS = ["exc", "cause", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, exc, cause, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.exc = exc
-        self.cause = cause
-
-
-class Try(stmt):
-    _FIELDS = [
-        "body",
-        "handlers",
-        "orelse",
-        "finalbody",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        body,
-        handlers,
-        orelse,
-        finalbody,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.body = body
-        self.handlers = handlers
-        self.orelse = orelse
-        self.finalbody = finalbody
-
-
-class Assert(stmt):
-    _FIELDS = ["test", "msg", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, test, msg, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.test = test
-        self.msg = msg
-
-
-class Import(stmt):
-    _FIELDS = ["names", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, names, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.names = names
-
-
-class ImportFrom(stmt):
-    _FIELDS = [
-        "module",
-        "names",
-        "level",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, module, names, level, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.module = module
-        self.names = names
-        self.level = level
-
-
-class Global(stmt):
-    _FIELDS = ["names", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, names, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.names = names
-
-
-class Nonlocal(stmt):
-    _FIELDS = ["names", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, names, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.names = names
-
-
-class Expr(stmt):
-    _FIELDS = ["value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-
-
-class Pass(stmt):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Break(stmt):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Continue(stmt):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class expr(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class BoolOp(expr):
-    _FIELDS = ["op", "values", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, op, values, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.op = op
-        self.values = values
-
-
-class BinOp(expr):
-    _FIELDS = [
-        "left",
-        "op",
-        "right",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, left, op, right, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.left = left
-        self.op = op
-        self.right = right
-
-
-class UnaryOp(expr):
-    _FIELDS = ["op", "operand", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, op, operand, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.op = op
-        self.operand = operand
-
-
-class Lambda(expr):
-    _FIELDS = ["args", "body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, args, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.args = args
-        self.body = body
-
-
-class IfExp(expr):
-    _FIELDS = [
-        "test",
-        "body",
-        "orelse",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, test, body, orelse, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.test = test
-        self.body = body
-        self.orelse = orelse
-
-
-class Dict(expr):
-    _FIELDS = ["keys", "values", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, keys, values, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.keys = keys
-        self.values = values
-
-
-class Set(expr):
-    _FIELDS = ["elts", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, elts, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.elts = elts
-
-
-class ListComp(expr):
-    _FIELDS = [
-        "elt",
-        "generators",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, elt, generators, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.elt = elt
-        self.generators = generators
-
-
-class SetComp(expr):
-    _FIELDS = [
-        "elt",
-        "generators",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, elt, generators, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.elt = elt
-        self.generators = generators
-
-
-class DictComp(expr):
-    _FIELDS = [
-        "key",
-        "value",
-        "generators",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, key, value, generators, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.key = key
-        self.value = value
-        self.generators = generators
-
-
-class GeneratorExp(expr):
-    _FIELDS = [
-        "elt",
-        "generators",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, elt, generators, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.elt = elt
-        self.generators = generators
-
-
-class Await(expr):
-    _FIELDS = ["value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-
-
-class Yield(expr):
-    _FIELDS = ["value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-
-
-class YieldFrom(expr):
-    _FIELDS = ["value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-
-
-class Compare(expr):
-    _FIELDS = [
-        "left",
-        "ops",
-        "comparators",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, left, ops, comparators, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.left = left
-        self.ops = ops
-        self.comparators = comparators
-
-
-class Call(expr):
-    _FIELDS = [
-        "func",
-        "args",
-        "keywords",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, func, args, keywords, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.func = func
-        self.args = args
-        self.keywords = keywords
-
-
-class FormattedValue(expr):
-    _FIELDS = [
-        "value",
-        "conversion",
-        "format_spec",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        value,
-        conversion,
-        format_spec,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-        self.conversion = conversion
-        self.format_spec = format_spec
-
-
-class JoinedStr(expr):
-    _FIELDS = ["values", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, values, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.values = values
-
-
-class Constant(expr):
-    _FIELDS = [
-        "value",
-        "kind",
-        "s",
-        "n",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, value, kind, s, n, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-        self.kind = kind
-        self.s = s
-        self.n = n
-
-
-class NamedExpr(expr):
-    _FIELDS = [
-        "target",
-        "value",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, target, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.target = target
-        self.value = value
-
-
-class Attribute(expr):
-    _FIELDS = [
-        "value",
-        "attr",
-        "ctx",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, value, attr, ctx, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-        self.attr = attr
-        self.ctx = ctx
-
-
-class slice(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Slice(slice):
-    _FIELDS = [
-        "lower",
-        "upper",
-        "step",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, lower, upper, step, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.lower = lower
-        self.upper = upper
-        self.step = step
-
-
-class ExtSlice(slice):
-    _FIELDS = ["dims", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, dims, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.dims = dims
-
-
-class Index(slice):
-    _FIELDS = ["value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-
-
-class Subscript(expr):
-    _FIELDS = [
-        "value",
-        "slice",
-        "ctx",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, value, slice, ctx, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-        self.slice = slice
-        self.ctx = ctx
-
-
-class Starred(expr):
-    _FIELDS = ["value", "ctx", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, value, ctx, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.value = value
-        self.ctx = ctx
-
-
-class Name(expr):
-    _FIELDS = ["id", "ctx", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, id, ctx, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.id = id
-        self.ctx = ctx
-
-
-class List(expr):
-    _FIELDS = ["elts", "ctx", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, elts, ctx, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.elts = elts
-        self.ctx = ctx
-
-
-class Tuple(expr):
-    _FIELDS = ["elts", "ctx", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, elts, ctx, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.elts = elts
-        self.ctx = ctx
-
-
-class expr_context(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class AugLoad(expr_context):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class AugStore(expr_context):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Param(expr_context):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Suite(mod):
-    _FIELDS = ["body", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, body, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.body = body
-
-
-class Del(expr_context):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Load(expr_context):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Store(expr_context):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class boolop(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class And(boolop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Or(boolop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class operator(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Add(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class BitAnd(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class BitOr(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class BitXor(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Div(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class FloorDiv(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class LShift(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Mod(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Mult(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class MatMult(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Pow(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class RShift(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Sub(operator):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class unaryop(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Invert(unaryop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Not(unaryop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class UAdd(unaryop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class USub(unaryop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class cmpop(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Eq(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Gt(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class GtE(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class In(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Is(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class IsNot(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class Lt(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class LtE(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class NotEq(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class NotIn(cmpop):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class comprehension(AST):
-    _FIELDS = [
-        "target",
-        "iter",
-        "ifs",
-        "is_async",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        target,
-        iter,
-        ifs,
-        is_async,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.target = target
-        self.iter = iter
-        self.ifs = ifs
-        self.is_async = is_async
-
-
-class excepthandler(AST):
-    _FIELDS = ["lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-
-
-class ExceptHandler(excepthandler):
-    _FIELDS = [
-        "type",
-        "name",
-        "body",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self, type, name, body, lineno, col_offset, end_lineno, end_col_offset
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.type = type
-        self.name = name
-        self.body = body
-
-
-class arguments(AST):
-    _FIELDS = [
-        "args",
-        "vararg",
-        "kwonlyargs",
-        "kw_defaults",
-        "kwarg",
-        "defaults",
-        "posonlyargs",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        args,
-        vararg,
-        kwonlyargs,
-        kw_defaults,
-        kwarg,
-        defaults,
-        posonlyargs,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.args = args
-        self.vararg = vararg
-        self.kwonlyargs = kwonlyargs
-        self.kw_defaults = kw_defaults
-        self.kwarg = kwarg
-        self.defaults = defaults
-        self.posonlyargs = posonlyargs
-
-
-class arg(AST):
-    _FIELDS = [
-        "arg",
-        "annotation",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(self, arg, annotation, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.arg = arg
-        self.annotation = annotation
-
-
-class keyword(AST):
-    _FIELDS = ["arg", "value", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, arg, value, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.arg = arg
-        self.value = value
-
-
-class alias(AST):
-    _FIELDS = ["name", "asname", "lineno", "col_offset", "end_lineno", "end_col_offset"]
-
-    def __init__(self, name, asname, lineno, col_offset, end_lineno, end_col_offset):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.name = name
-        self.asname = asname
-
-
-class withitem(AST):
-    _FIELDS = [
-        "context_expr",
-        "optional_vars",
-        "lineno",
-        "col_offset",
-        "end_lineno",
-        "end_col_offset",
-    ]
-
-    def __init__(
-        self,
-        context_expr,
-        optional_vars,
-        lineno,
-        col_offset,
-        end_lineno,
-        end_col_offset,
-    ):
-        super().__init__(lineno, col_offset, end_lineno, end_col_offset)
-        self.context_expr = context_expr
-        self.optional_vars = optional_vars
+import ast
+import inspect
+import typing
+from collections import defaultdict
+
+from . import doc_core as doc
+from .doc_core import *  # pylint: disable=unused-import,wildcard-import,redefined-builtin,W0614
+
+FnToDoc = typing.Callable[[ast.AST], doc.AST]
+FnFromDoc = typing.Callable[[doc.AST], ast.AST]
+
+
+class Entry:
+    to_doc: typing.Optional[FnToDoc]
+    from_doc: typing.Optional[FnFromDoc]
+
+    def __init__(self):
+        self.to_doc = None
+        self.from_doc = None
+
+
+class Registry:
+    _inst: typing.Optional["Registry"] = None
+    table: typing.Dict[str, Entry]
+
+    def __init__(self):
+        self.table = defaultdict(Entry)
+
+
+def register_to_doc(name: str):
+    def f(to_doc: FnToDoc):  # pylint: disable=redefined-outer-name
+        reg = Registry._inst  # pylint: disable=protected-access
+        reg.table[name].to_doc = to_doc
+
+    return f
+
+
+def register_from_doc(name: str):
+    def f(to_doc: FnFromDoc):  # pylint: disable=redefined-outer-name
+        reg = Registry._inst  # pylint: disable=protected-access
+        reg.table[name].from_doc = to_doc
+
+    return f
+
+
+def _is_atomic_type(node):
+    return (
+        node is None
+        or node in [..., True, False]
+        or isinstance(
+            node,
+            (
+                int,
+                float,
+                str,
+                bool,
+                bytes,
+                complex,
+            ),
+        )
+    )
+
+
+def _get_registry_entry(cls_name, attr):
+    cls_name = cls_name.split(".")[-1]
+    reg = Registry._inst  # pylint: disable=protected-access
+    if cls_name in reg.table:
+        entry = reg.table[cls_name]
+        return getattr(entry, attr, None)
+    return None
+
+
+def from_doc(node):
+    if _is_atomic_type(node):
+        return node
+    if isinstance(node, tuple):
+        return tuple(from_doc(n) for n in node)
+    if isinstance(node, list):
+        return [from_doc(n) for n in node]
+    func = _get_registry_entry(node.__class__.__name__, "from_doc")
+    if not func:
+        raise NotImplementedError(f"from_doc is not implemented for: {node.__class__.__name__}")
+    return func(node)
+
+
+def to_doc(node):
+    if _is_atomic_type(node):
+        return node
+    if isinstance(node, tuple):
+        return tuple(to_doc(n) for n in node)
+    if isinstance(node, list):
+        return [to_doc(n) for n in node]
+    func = _get_registry_entry(node.__class__.__name__, "to_doc")
+    if not func:
+        raise NotImplementedError(f"to_doc is not implemented for: {node.__class__.__name__}")
+    return func(node)
+
+
+def _register_default():
+    class DefaultTranslator:
+        def __init__(self, doc_cls, func, fields):
+            self.doc_cls = doc_cls  # getattr(doc, name)
+            self.func = func
+            self.fields = fields
+
+        def __call__(self, node):
+            kv = {attr: self.func(getattr(node, attr, None)) for attr in self.fields}
+            return self.doc_cls(**kv)
+
+    Registry._inst = Registry()  # pylint: disable=protected-access
+    for cls_name in dir(doc):
+        doc_cls = getattr(doc, cls_name)
+        if inspect.isclass(doc_cls) and issubclass(doc_cls, doc.AST):
+            assert "." not in cls_name
+            register_to_doc(cls_name)(
+                DefaultTranslator(
+                    getattr(doc, cls_name),
+                    to_doc,
+                    doc_cls._FIELDS,  # pylint: disable=protected-access
+                )
+            )
+            register_from_doc(cls_name)(
+                DefaultTranslator(
+                    getattr(ast, cls_name),
+                    from_doc,
+                    doc_cls._FIELDS,  # pylint: disable=protected-access
+                )
+            )
+
+
+def parse(
+    source,
+    filename="<unknown>",
+    mode="exec",
+    *,
+    type_comments=False,
+    feature_version=None,
+) -> doc.AST:
+    program = ast.parse(
+        source=source,
+        filename=filename,
+        mode=mode,
+        type_comments=type_comments,
+        feature_version=feature_version,
+    )
+    return to_doc(program)
+
+
+class NodeVisitor:
+    def visit(self, node: doc.AST) -> None:
+        if isinstance(node, (list, tuple)):
+            for item in node:
+                self.visit(item)
+        return getattr(
+            self,
+            "visit_" + node.__class__.__name__.split(".")[-1],
+            self.generic_visit,
+        )(node)
+
+    def generic_visit(self, node: doc.AST) -> None:
+        for field in node.__class__._FIELDS:  # pylint: disable=protected-access
+            value = getattr(node, field, None)
+            if value is None:
+                pass
+            elif isinstance(value, doc.AST):
+                self.visit(value)
+            elif isinstance(value, (list, tuple)):
+                for item in value:
+                    if isinstance(item, doc.AST):
+                        self.visit(item)
+
+
+_register_default()
